+++ tsconfig.app.json
@@ -1,6 +1,6 @@
 {
   "extends": "@vue/tsconfig/tsconfig.dom.json",
-  "include": ["env.d.ts", "src/**/*", "src/**/*.json", "src/**/*.ts", "src/**/*.vue"],
+  "include": ["env.d.ts", "src/**/*", "src/**/*.json", "src/**/*.ts", "src/**/**/*.vue"],
   "exclude": ["src/**/__tests__/*"],
   "compilerOptions": {
     "composite": true,
